requirements.txt
streamlit
opencv-python
numpy
Pillow
torch
torchvision